backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
GeoAlchemy2==0.17.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
packaging==24.2
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0
